requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
chromadb==0.4.22
openai==1.3.7
google-generativeai==0.3.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-google-genai==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
tiktoken==0.5.2
typing-extensions==4.8.0

